requirements.txt
# X-Stock Agent Dependencies

# Core Framework
vnpy==3.9.0
akshare>=1.12.0
tushare>=1.2.89

# Data & Cache
sqlite3
redis>=4.5.0
pandas>=2.0.0
numpy>=1.24.0

# Web Dashboard
streamlit>=1.30.0
plotly>=5.18.0
altair>=5.2.0

# Analysis & ML
scikit-learn>=1.3.0
torch>=2.1.0
ta-lib>=0.4.28

# Scheduling & Automation
apscheduler>=3.10.0
schedule>=1.2.0

# Utilities
pyyaml>=6.0.1
requests>=2.31.0
loguru>=0.7.2
python-dotenv>=1.0.0

# Sentiment Analysis
jieba>=0.42.1
snowNLP>=0.12.3

# Testing
pytest>=7.4.0
